+++ packages/orq-rc/RELEASES.md
@@ -9019,3 +9019,13 @@ Based on:
 - [typescript v4.9.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.40 - packages/orq-rc
+
+## 2026-05-11 12:09:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.41 - packages/orq-rc